requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.3
opencv-python==4.12.0.88
ultralytics==8.3.187
supervision==0.26.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
numpy==2.2.6
